basic specs

+++ lib/smusher.rb
@@ -6,7 +6,29 @@ class Smusher
   SMUSHIT_FAILURE_SIZE = 9667  
   
   def store_smushed_image(url,file)
+    url = sanitize_url(url)
+    with_protection(file) do
+      with_logging(url,file) do
+        write_smushed_data(url,file)
+      end
+    end
+  end
+
+  def store_smushed_folder(base_url,folder)
+    base_url = sanitize_url(base_url)
+    folder = sanitize_folder(folder)
     
+    images_in_folder(folder).each do |file|
+      relative_path_to_folder = file.sub(folder+'/','')
+      url = "#{base_url}/#{relative_path_to_folder}"
+      store_smushed_image(url,file)
+      puts ''
+    end
+  end
+
+private
+
+  def write_smushed_data(url,file)
     File.open(file,'w') do |f|
       data = smushed_image_data_for(url)
       raise data if data =~ /<html>/i
@@ -14,27 +36,14 @@ class Smusher
     end
   end
 
-  def store_smushed_folder(url,folder)
-    url = url.dup#unfreeze
-    sanitize!(folder,url)
-    
-    images_in_folder(folder).each do |file|
-      relative_path_to_folder = file.sub(folder+'/','')
-      online_file = "#{url}/#{relative_path_to_folder}"
-      with_protection(file) do
-        with_logging(online_file,file) do
-          store_smushed_image(online_file,file)
-        end
-      end
-    end
+  def sanitize_url(url)
+    url = url.sub(%r[/$],'')#remove tailing slash
+    url = url.sub(url,"http://#{url}") unless url.include?('://')#add http if protocol is missing
+    url
   end
-
-private
-
-  def sanitize!(folder,url)
-    folder.sub!(%r[/$],'')#remove tailing slash
-    url.sub!(%r[/$],'')#remove tailing slash
-    url.sub!(url,"http://#{url}") unless url.include?('://')#add http if protocol is missing
+  
+  def sanitize_folder(folder)
+    folder.sub(%r[/$],'')#remove tailing slash
   end
 
   def images_in_folder(folder)
@@ -43,6 +52,8 @@ private
   end
   
   def with_protection(file)
+    return yield unless File.exist?(file)
+    
     backup = "#{file}.backup"
     FileUtils.cp(file,backup)#backup
     
@@ -64,12 +75,13 @@ private
 
   def with_logging(url,file)
     puts "sushing #{url} -> #{file}"
+    
     before = size(file)
     yield(file)
     after = size(file)
+    
     result = before == 0 ?  "CREATED" : "#{(100*after)/before}%"
     puts "#{before} -> #{after}".ljust(40) + " = #{result}"
-    puts ''
   end
 
   def smushed_image_data_for(url)
